+++ torchstain/__init__.py
@@ -1,3 +1,4 @@
 __version__ = '1.1.0'
 
 from torchstain.base import normalizers
+from torchstain import torch, tf
