+++ torchstain/base/normalizers/macenko.py
@@ -3,10 +3,10 @@ def MacenkoNormalizer(backend='torch'):
         from torchstain.numpy.normalizers import NumpyMacenkoNormalizer
         return NumpyMacenkoNormalizer()
     elif backend == "torch":
-        from torchstain.torch_backend.normalizers.macenko import TorchMacenkoNormalizer
+        from torchstain.torch.normalizers.macenko import TorchMacenkoNormalizer
         return TorchMacenkoNormalizer()
     elif backend == "tensorflow":
-        from torchstain.tf_backend.normalizers.macenko import TensorFlowMacenkoNormalizer
+        from torchstain.tf.normalizers.macenko import TensorFlowMacenkoNormalizer
         return TensorFlowMacenkoNormalizer()
     else:
         raise Exception(f'Unknown backend {backend}')
