+++ torchstain/tf/normalizers/__init__.py
@@ -1,0 +1,1 @@
+from torchstain.tf.normalizers.macenko import TensorFlowMacenkoNormalizer
