--- torchstain/tf_backend/normalizers/__init__.py
@@ -1,1 +1,0 @@
-from torchstain.tf_backend.normalizers.macenko import TensorFlowMacenkoNormalizer
