--- torchstain/tf_backend/utils/__init__.py
@@ -1,3 +1,0 @@
-from torchstain.tf_backend.utils.cov import cov_tf
-from torchstain.tf_backend.utils.percentile import percentile_tf
-from torchstain.tf_backend.normalizers.solveLS import solveLS
