+++ torchstain/torch/normalizers/__init__.py
@@ -1,0 +1,1 @@
+from torchstain.torch.normalizers.macenko import TorchMacenkoNormalizer
