+++ torchstain/torch/utils/__init__.py
@@ -1,0 +1,2 @@
+from torchstain.torch.utils.cov import cov
+from torchstain.torch.utils.percentile import percentile
