--- torchstain/torch_backend/normalizers/__init__.py
@@ -1,1 +1,0 @@
-from torchstain.torch_backend.normalizers.macenko import TorchMacenkoNormalizer
